+++ tests/testthat/test-spec2Munsell.R
@@ -27,10 +27,50 @@ test_that("spec2Munsell works as expected", {
     )
   
   # expected dE00
-  expect_equal(m$sigma, 0.64280, tolerance = 0.01)
+  expect_equal(m$sigma, 0.64, tolerance = 0.01)
       
 })
 
+
+test_that("interpolated / extrapolated spectra", {
+  
+  # 1-chroma chips are extrapolated for some hues
+  # depends on SO and illuminant
+  chip <- '7.5YR 2/1'
+  m <- spec2Munsell(munsell.spectra.wide[, chip], SO = 'CIE1931', illuminant = 'D65')  
+  
+  # object structure / contents
+  expect_true(inherits(m, 'data.frame'))
+  
+  # same output as col2Munsell
+  expect_true(
+    all(
+      names(m) == c('hue', 'value', 'chroma', 'sigma')
+    )
+  )
+  
+  # input == output
+  expect_true(
+    sprintf("%s %s/%s", m$hue, m$value, m$chroma) == chip
+  )
+  
+  # expected dE00
+  expect_equal(m$sigma, 3, tolerance = 0.1)
+  
+  
+  # 8.5-value chips are interpolated for some hues
+  chip <- '7.5YR 8.5/6'
+  m <- spec2Munsell(munsell.spectra.wide[, chip], SO = 'CIE1931', illuminant = 'D65') 
+  
+  # input == output
+  expect_true(
+    sprintf("%s %s/%s", m$hue, m$value, m$chroma) == chip
+  )
+  
+  # expected dE00
+  expect_equal(m$sigma, 0.13, tolerance = 0.01)
+  
+})
 
 
 test_that("supporting data match range of Munsell spectra", {